Update Santa recipes (#62)
* Update Santa recipes

This PR updates the Santa recipes to streamline a few things and fix several problems. It partially resolves #57.

### santa.download.recipe

- Downloads the package directly instead of the DMG (which contains the same package)

### santa.install.recipe

- Updates the path to point directly to the downloaded package

### santa.munki.recipe

- Converts the `installcheck_script` from python to Zshell now that Munki Python is deprecated.

This script now checks to make sure Santa.app is installed, is the right version and uses `santactl` to evaluate if things are running correctly.

- Switches the `uninstall_method` and provides an `uninstall_script` sourced from the project's GitHub repo

The existing recipe defaults to `removepackages` as the uninstall method which is not effective since the package deploys to /tmp and then moves files into place with a postinstall script.

- Fixes the installs array generation which does not work at all in the existing recipe
- Removes a hardcoded `minimum_os_version` and replaces it with automatic detection as part of the `MunkiInstallsItemsCreator` processor
- Changes the `Versioner` logic to use `CFBundleVersion` which the current recipe attempts to use in the `MunkiImporter` but doesn't work

This also projects against the possibility of two releases in a single month.

- Adds a `category` to the input
- Adds a `developer` to the input
- Adds `PathDeleter` processor to clean up directories

* Simplify version check

* Remove extra line break

* Update santa.munki.recipe

Fix variable